[Refactor][MRV2] Share KV cache binding helpers
Move common cache binding into a shared worker helper used by MRV1 and MRV2. Preserve DeepSeek-V4 V1 model/MTP ordering and virtual-engine wrapping in a dedicated path, and remove the redundant Qwen3 patch.
