+++ vllm_ascend/patch/worker/patch_v2/patch_attn_utils.py
@@ -1,5 +1,6 @@
 import vllm
 
+from vllm_ascend.patch.worker.patch_bind_kv_cache import bind_kv_cache
 from vllm_ascend.worker.v2.attn_utils import (
     _allocate_kv_cache,
     _reshape_kv_cache_v2,
@@ -8,4 +9,5 @@ from vllm_ascend.worker.v2.attn_utils import (
 
 vllm.v1.worker.gpu.attn_utils._allocate_kv_cache = _allocate_kv_cache
 vllm.v1.worker.gpu.attn_utils._reshape_kv_cache = _reshape_kv_cache_v2
+vllm.v1.worker.gpu.attn_utils.bind_kv_cache = bind_kv_cache
 vllm.v1.worker.gpu.model_runner.get_kv_cache_spec = get_kv_cache_spec
